Move the coverage based strategy to the strategies packet and write documentation on the methods
