--- pkg/strategies/coveragebased/BUILD
@@ -1,49 +1,0 @@
-# Copyright 2023 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-package(
-    default_visibility = [
-        "//visibility:public",
-    ],
-)
-
-go_library(
-    name = "coveragebased",
-    srcs = [
-        "coverage_based.go",
-        "heap.go",
-    ],
-    importpath = "buzzer/pkg/strategies/coveragebased/coveragebased",
-    deps = [
-        "//pkg/ebpf",
-        "//pkg/rand",
-        "//pkg/strategies",
-        "//pkg/units",
-        "//proto:ebpf_go_proto",
-        "//proto:ffi_go_proto",
-        "@com_github_golang_protobuf//proto",
-    ],
-)
-
-go_test(
-    name = "coveragebased_test",
-    srcs = [
-        "heap_test.go",
-    ],
-    embed = [":coveragebased"],
-    importpath = "buzzer/pkg/strategies/coveragebased/coveragebased",
-    deps = [],
-)
